% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{lauch_app}
\alias{lauch_app}
\title{Launch shiny app for visualising pedestrian and weather data}
\usage{
lauch_app()
}
\description{
Launch shiny app for visualising pedestrian and weather data
}
\examples{
\dontrun{
launchApp()
}
}
