% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{launch_app}
\alias{launch_app}
\title{Launch shiny app for visualising pedestrian and weather data}
\usage{
launch_app()
}
\description{
Launch shiny app for visualising pedestrian and weather data
}
\examples{
\dontrun{
   launch_app()
}
}
