% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{melb_prcp}
\alias{melb_prcp}
\title{Monthly cumulative precipitation recorded in Melbourne}
\format{A data frame with 366 rows and 2 variables:
\describe{
  \item{date}{Date of 2016}
  \item{prcp}{Monthly cumulative precipitation}
}}
\usage{
melb_prcp
}
\description{
A dataset containing daily precipitations in Melbourne 2016. The variables 
are as follows:
}
\examples{
melb_prcp
}
